LPD-68389 Rename ContentEditorManagementBar to ContentEditorToolbar
