+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layout-page-templates/master-pages/cms-translation-master/page-definition.json
@@ -13,7 +13,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ContentEditorManagementBarComponentSectionFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ContentEditorToolbarComponentSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
