+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layout-page-templates/master-pages/content-editor-master/page-definition.json
@@ -27,7 +27,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ContentEditorManagementBarComponentSectionFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ContentEditorToolbarComponentSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
