+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/css/content_editor/_variables.scss
@@ -1,6 +1,6 @@
-$contentEditorManagementBarHeight: 64px;
 $contentEditorSidePanelBarWidth: 40px;
 $contentEditorSidePanelWidth: 280px;
+$contentEditorToolbarHeight: 64px;
 
 $pageEditorManagamentBarHeight: 64px;
 $pageEditorSidebarButtonsWidth: 42px;
